1、增加修改Jwt相关内容，并增加JwtUser类 2、修改DateTimeKit 3、BaseController增加获取JwtUser对象函数

+++ ktf-common/src/main/java/com/kivi/framework/util/kit/DateTimeKit.java
@@ -128,8 +128,8 @@ public class DateTimeKit {
      * 
      * @return
      */
-    public static long currentUnix() {
-        return System.currentTimeMillis() / 1000;
+    public static int currentUnix() {
+        return (int) (System.currentTimeMillis() / 1000);
     }
 
     /**
